feat: add property and task types; implement new screens for cleaner and house owner; refactor layout and imports
