+++ app/(pages)/(account)/_layout.tsx
@@ -8,9 +8,9 @@ export default function RootLayout() {
 function RootLayoutNav() {
   return (
     <Stack screenOptions={{ headerShown: false }}>
-      <Stack.Screen name="notification" options={{ headerShown: false }} />
+      <Stack.Screen name="notifications" options={{ headerShown: false }} />
       <Stack.Screen name="settings" options={{ headerShown: false }} />
-      <Stack.Screen name="payment" options={{ headerShown: false }} />
+      <Stack.Screen name="(payment)" options={{ headerShown: false }} />
     </Stack>
   );
 }
