+++ app/(pages)/(profile)/editProfile.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import {
   View,
   Text,
@@ -11,7 +11,7 @@ import {
 import { useRouter } from 'expo-router';
 import { Ionicons } from '@expo/vector-icons';
 import { supabase } from '@/clients/supabase';
-import { useStore } from 'zustand';
+import useStore from '@/lib/store';
 
 export default function EditProfile() {
   const router = useRouter();
