--- styles/nextButton.js
@@ -1,28 +1,0 @@
-// CustomButton.js
-import React from "react";
-import { TouchableOpacity, Text, StyleSheet } from "react-native";
-
-const NextButton = ({ title, onPress, style }) => {
-  return (
-    <TouchableOpacity style={[styles.button, style]} onPress={onPress}>
-      <Text style={styles.buttonText}>{title}</Text>
-    </TouchableOpacity>
-  );
-};
-
-const styles = StyleSheet.create({
-  button: {
-    backgroundColor: "#4E89CE",
-    padding: 15,
-    borderRadius: 8,
-    alignItems: "center",
-    marginTop: 10,
-  },
-  buttonText: {
-    color: "#fff",
-    fontSize: 16,
-    fontWeight: "bold",
-  },
-});
-
-export default NextButton;
